work in progress... draft postion closing
